refactor(CLOUD-234): moved the resource code to the appropriate files
